+++ openbank-flaky-test-hunter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.14.3](https://github.com/JiRaska/open-bank-oss/compare/flaky-test-hunter-v0.14.2...flaky-test-hunter-v0.14.3) (2026-09-01)
+
+
+### Bug Fixes
+
+* **flaky-test-hunter:** reject a null array element with 400 instead of 500 ([#7994](https://github.com/JiRaska/open-bank-oss/issues/7994)) ([5c29732](https://github.com/JiRaska/open-bank-oss/commit/5c29732654d46781e08be0fa853028a8b0a37fb9))
+
 ## [0.14.2](https://github.com/JiRaska/open-bank-oss/compare/flaky-test-hunter-v0.14.1...flaky-test-hunter-v0.14.2) (2026-08-30)